move JvmOtlpRuntimeMetrics.java to agent-jmxfetch
