+++ dd-java-agent/agent-bootstrap/src/main/resources/META-INF/native-image/com.datadoghq/dd-java-agent/reflect-config.json
@@ -218,12 +218,5 @@
     "methods": [
       {"name": "<init>", "parameterTypes": []}
     ]
-  },
-  {
-    "name": "datadog.trace.bootstrap.otel.shim.metrics.JvmOtlpRuntimeMetrics",
-    "methods": [
-      {"name": "<init>", "parameterTypes": []},
-      {"name": "start", "parameterTypes": []}
-    ]
   }
 ]
